Removing old implementation
