+++ gemsfx-demo/src/main/java/com/dlsc/gemsfx/demo/PhoneNumberFieldApp.java
@@ -1,6 +1,10 @@
 package com.dlsc.gemsfx.demo;
 
 import com.dlsc.gemsfx.PhoneNumberField;
+import com.google.i18n.phonenumbers.AsYouTypeFormatter;
+import com.google.i18n.phonenumbers.NumberParseException;
+import com.google.i18n.phonenumbers.PhoneNumberUtil;
+import com.google.i18n.phonenumbers.Phonenumber;
 import javafx.application.Application;
 import javafx.beans.InvalidationListener;
 import javafx.beans.binding.Bindings;
@@ -9,8 +13,8 @@ import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
-import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
 import javafx.scene.layout.HBox;
@@ -28,35 +32,62 @@ public class PhoneNumberFieldApp extends Application {
             return null;
         }
         PhoneNumberField.CountryCallingCode code = (PhoneNumberField.CountryCallingCode) c;
-        return "(+" + code.countryCode() + ") " + code;
+        return "(" + code.phonePrefix() + ") " + code;
     };
 
     @Override
     public void start(Stage stage) throws Exception {
+        PhoneNumberUtil phoneUtil = PhoneNumberUtil.getInstance();
+
         PhoneNumberField field = new PhoneNumberField();
+
+        field.setPhoneNumberValidator(phoneNumber -> {
+            if (field.getCountryCallingCode() == null ||
+                phoneNumber == null || phoneNumber.isEmpty() ||
+                field.getLocalPhoneNumber() == null || field.getLocalPhoneNumber().isEmpty()) {
+                return true;
+            }
+
+            try {
+                Phonenumber.PhoneNumber number = phoneUtil.parse(phoneNumber, field.getCountryCallingCode().iso2Code());
+                return phoneUtil.isValidNumber(number);
+            } catch (NumberParseException e) {
+                System.err.println(e.getLocalizedMessage());
+            }
+
+            return false;
+        });
+
+        field.setPhoneNumberFormatter(phoneNumber -> {
+            if (phoneNumber == null || phoneNumber.isEmpty() || field.getCountryCallingCode() == null) {
+                return "";
+            }
+
+            AsYouTypeFormatter formatter = phoneUtil.getAsYouTypeFormatter(field.getCountryCallingCode().iso2Code());
+            String formatted = "";
+            for (char c : phoneNumber.toCharArray()) {
+                formatted = formatter.inputDigit(c);
+            }
+            return formatted;
+        });
 
         VBox controls = new VBox(10);
         addControl("Available Countries", availableCountriesSelector(field), controls);
         addControl("Preferred Countries", preferredCountriesSelector(field), controls);
-        addControl("Default Country", defaultCountrySelector(field), controls);
         addControl("Disable Country", disableCountryCheck(field), controls);
-        addControl("Force Local Phone", forceLocalPhoneNumberCheck(field), controls);
-        addControl("Strict Mode", strictModeCheck(field), controls);
-        addControl("Unmasked", unmaskedModeCheck(field), controls);
+        addControl("", clearButton(field), controls);
 
         VBox fields = new VBox(10);
         addField(fields, "Country Code", field.countryCallingCodeProperty(), COUNTRY_CODE_CONVERTER);
         addField(fields, "Phone Number", field.phoneNumberProperty());
         addField(fields, "Local Number", field.localPhoneNumberProperty());
-        addField(fields, "Formatted Local Number", field.formattedLocalPhoneNumberProperty());
-        addField(fields, "Mask", field.maskProperty());
 
         VBox vBox = new VBox(20);
         vBox.setPadding(new Insets(20));
         vBox.setAlignment(Pos.CENTER);
         vBox.getChildren().addAll(controls, new Separator(), field, new Separator(), fields);
 
-        stage.setTitle("PhoneNumberField");
+        stage.setTitle("PhoneNumberField2");
         stage.setScene(new Scene(vBox, 500, 500));
         stage.sizeToScene();
         stage.centerOnScreen();
@@ -98,42 +129,16 @@ public class PhoneNumberFieldApp extends Application {
         return comboBox;
     }
 
-    private Node defaultCountrySelector(PhoneNumberField view) {
-        ComboBox<PhoneNumberField.CountryCallingCode> comboBox = new ComboBox<>();
-        comboBox.getItems().add(null);
-        comboBox.getItems().addAll(PhoneNumberField.CountryCallingCode.Defaults.values());
-        comboBox.valueProperty().bindBidirectional(view.countryCallingCodeProperty());
-        return comboBox;
-    }
-
     private Node disableCountryCheck(PhoneNumberField field) {
         CheckBox check = new CheckBox();
         check.selectedProperty().bindBidirectional(field.disableCountryCodeProperty());
         return check;
     }
 
-    private Node forceLocalPhoneNumberCheck(PhoneNumberField field) {
-        CheckBox check = new CheckBox();
-        check.selectedProperty().bindBidirectional(field.forceLocalNumberProperty());
-        return check;
-    }
-
-    private Node strictModeCheck(PhoneNumberField field) {
-        CheckBox check = new CheckBox();
-        check.selectedProperty().bindBidirectional(field.strictModeProperty());
-        return check;
-    }
-
-    private Node unmaskedModeCheck(PhoneNumberField field) {
-        CheckBox check = new CheckBox();
-        check.selectedProperty().addListener((obs, oldV, newV) -> {
-            if (newV) {
-                field.setMaskProvider(null);
-            } else {
-                field.setMaskProvider(PhoneNumberField.DEFAULT_MASK_PROVIDER);
-            }
-        });
-        return check;
+    private Node clearButton(PhoneNumberField field) {
+        Button clear = new Button("Clear all");
+        clear.setOnAction(evt -> field.setPhoneNumber(null));
+        return clear;
     }
 
     private void addControl(String name, Node control, VBox controls) {
