+++ gemsfx/src/main/java/com/dlsc/gemsfx/PhoneNumberField.java
@@ -4,6 +4,7 @@ import com.dlsc.gemsfx.skins.PhoneNumberFieldSkin;
 import javafx.application.Platform;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.ReadOnlyBooleanWrapper;
 import javafx.beans.property.ReadOnlyStringProperty;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.beans.property.SimpleBooleanProperty;
@@ -12,15 +13,17 @@ import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.css.PseudoClass;
 import javafx.scene.Node;
 import javafx.scene.control.Control;
 import javafx.scene.control.Skin;
 import javafx.scene.control.TextField;
 import javafx.scene.control.TextFormatter;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
 import javafx.util.Callback;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -31,91 +34,56 @@ import java.util.function.UnaryOperator;
 
 /**
  * <p>A control for entering phone numbers. By default, the phone numbers are expressed in international format and will be
- * delivered via the {@link #phoneNumberProperty() phone number} property, however, it would also be possible to enter local phone number by
- * using {@link #forceLocalNumberProperty() force local} property.</p>
- *
- * <p>A masking technique was implemented to let the user know what is the
- * expected format, you can control it by using a custom {@link #maskProviderProperty()}.</p>
+ * delivered via the {@link #phoneNumberProperty() phone number} property.</p>
  *
  * <p>The control supports a list of {@link #getAvailableCountryCodes() available country codes} and works based on
  * the {@link CountryCallingCode CountryCallingCode} interface. This interface allows customizing the country codes and their
  * respective area codes, in case the default values are not sufficient or the numbers change.
  * </p>
+ *
+ * <p>Formatting of the phone number is possible via the {@link #phoneNumberFormatterProperty() phoneNumberFormatter} property.
+ * </p>
  */
 public class PhoneNumberField extends Control {
+
+    public static final PseudoClass ERROR_PSEUDO_CLASS = PseudoClass.getPseudoClass("error");
 
     /**
      * Default style class for css styling.
      */
     public static final String DEFAULT_STYLE_CLASS = "phone-number-field";
 
-    /**
-     * Default mask for the phone number field.
-     */
-    public static final String DEFAULT_MASK = "(___) ___-____";
-
-    /**
-     * Default character used to represent a digit in the mask.
-     */
-    public static final char DEFAULT_MASK_DIGIT_CHAR = '_';
-
-    public static final Callback<PhoneNumberField, String> DEFAULT_MASK_PROVIDER = field -> {
-        if (field.isForceLocalNumber()) {
-            return DEFAULT_MASK;
-        }
-        if (field.getCountryCallingCode() == null) {
-            return "";
-        }
-        return field.getCountryCallingCode().mask(field.getLocalPhoneNumber());
-    };
-
-    /**
-     * List of supported characters a mask can have.
-     */
-    public static final List<Character> DEFAULT_MASK_SUPPORTED_CHARS = Arrays.asList(DEFAULT_MASK_DIGIT_CHAR, '-', '(', ')', ' ');
-
     private final PhoneNumberParser parser;
     private final PhoneNumberFormatter formatter;
+    private final TextField textField;
 
     /**
      * Builds a new phone number field with the default settings. The available country calling codes is taken from
      * {@link CountryCallingCode.Defaults Defaults}.
      */
     public PhoneNumberField() {
-        parser = new PhoneNumberParser();
-        formatter = new PhoneNumberFormatter();
         getStyleClass().add(DEFAULT_STYLE_CLASS);
         getAvailableCountryCodes().setAll(CountryCallingCode.Defaults.values());
 
-        Runnable maskUpdater = () -> setMask(Optional.ofNullable(getMaskProvider()).map(p -> p.call(PhoneNumberField.this)).orElse(""));
-        countryCallingCodeProperty().addListener((obs, oldV, newV) -> maskUpdater.run());
-        localPhoneNumberProperty().addListener((obs, oldV, newV) -> maskUpdater.run());
-        maskProviderProperty().addListener((obs, oldV, newV) -> maskUpdater.run());
-        setMaskProvider(DEFAULT_MASK_PROVIDER);
+        textField = new TextField();
 
-        Runnable maskListener = () -> {
-            String mask = getMask();
-            String formattedValue = getFormattedLocalPhoneNumber();
-            if (mask == null || mask.isEmpty()) {
-                setMaskRemaining("");
-            } else {
-                if (formattedValue == null || formattedValue.isEmpty()) {
-                    setMaskRemaining(mask);
-                } else if (formattedValue.length() > mask.length()) {
-                    setMaskRemaining("");
-                } else {
-                    setMaskRemaining(mask.substring(formattedValue.length()));
-                }
-            }
-        };
+        parser = new PhoneNumberParser();
+        formatter = new PhoneNumberFormatter(textField);
 
-        maskProperty().addListener((obs, oldV, newV) -> Platform.runLater(() -> {
-            // Needs to run after completed text input
-            formatter.setFormattedLocalPhoneNumber(getLocalPhoneNumber());
-            maskListener.run();
-        }));
 
-        formattedLocalPhoneNumberProperty().addListener((obs, oldV, newV) -> maskListener.run());
+        Runnable formatUpdater = () -> Platform.runLater(() -> formatter.setFormattedLocalPhoneNumber(getLocalPhoneNumber()));
+        phoneNumberFormatter.addListener((obs, oldV, newV) -> formatUpdater.run());
+
+        Runnable validUpdater = () -> Platform.runLater(() -> setValid(Optional.ofNullable(getPhoneNumberValidator()).map(v -> v.call(getPhoneNumber())).orElse(true)));
+        phoneNumberValidator.addListener((obs, oldV, newV) -> validUpdater.run());
+
+        phoneNumber.addListener((obs, oldV, newV) -> {
+            formatUpdater.run();
+            validUpdater.run();
+        });
+
+        valid.addListener((obs, oldV, newV) -> pseudoClassStateChanged(ERROR_PSEUDO_CLASS, !newV));
+        validUpdater.run();
     }
 
     @Override
@@ -125,7 +93,7 @@ public class PhoneNumberField extends Control {
 
     @Override
     protected Skin<?> createDefaultSkin() {
-        return new PhoneNumberFieldSkin(this, formatter.textField);
+        return new PhoneNumberFieldSkin(this, textField);
     }
 
     // VALUES
@@ -148,14 +116,6 @@ public class PhoneNumberField extends Control {
                 // Set the value first, so that the binding will be triggered
                 super.set(newPhoneNumber);
 
-                if (isForceLocalNumber()) {
-                    // No need to infer the country code, just use the local phone number
-                    setCountryCallingCode(null);
-                    setLocalPhoneNumber(newPhoneNumber);
-                    formatter.setFormattedLocalPhoneNumber(newPhoneNumber);
-                    return;
-                }
-
                 // Set depending fields here
                 PhoneNumber phoneNumber = parser.call(newPhoneNumber);
 
@@ -168,8 +128,7 @@ public class PhoneNumberField extends Control {
                     setLocalPhoneNumber(phoneNumber.localPhoneNumber);
                     formatter.setFormattedLocalPhoneNumber(phoneNumber.localPhoneNumber);
                 }
-            }
-            finally {
+            } finally {
                 selfUpdate = false;
             }
         }
@@ -206,10 +165,8 @@ public class PhoneNumberField extends Control {
                 super.set(newCountryCallingCode);
 
                 // For now replace the entire text, it might be good to preserve the local number and just change the country code
-                setPhoneNumber(isForceLocalNumber() ?  null :
-                    Optional.ofNullable(newCountryCallingCode)
+                setPhoneNumber(Optional.ofNullable(newCountryCallingCode)
                         .map(CountryCallingCode::phonePrefix)
-                        .map(String::valueOf)
                         .orElse(null));
 
             } finally {
@@ -249,19 +206,6 @@ public class PhoneNumberField extends Control {
 
     private void setLocalPhoneNumber(String localPhoneNumber) {
         this.localPhoneNumber.set(localPhoneNumber);
-    }
-
-    private final ReadOnlyStringWrapper formattedLocalPhoneNumber = new ReadOnlyStringWrapper(this, "formattedLocalPhoneNumber");
-
-    /**
-     * @return Read only property that exposes the local phone number expressed as the mask format defines.
-     */
-    public final ReadOnlyStringProperty formattedLocalPhoneNumberProperty() {
-        return formattedLocalPhoneNumber.getReadOnlyProperty();
-    }
-
-    public final String getFormattedLocalPhoneNumber() {
-        return formattedLocalPhoneNumber.get();
     }
 
     // SETTINGS
@@ -304,46 +248,6 @@ public class PhoneNumberField extends Control {
         disableCountryCodeProperty().set(disableCountryCode);
     }
 
-    private final BooleanProperty forceLocalNumber = new SimpleBooleanProperty(this, "forceLocalNumber") {
-        private CountryCallingCode lastCountryCallingCode;
-        @Override
-        public void set(boolean newForceLocal) {
-            super.set(newForceLocal);
-
-            final String localPhone = Optional.ofNullable(getLocalPhoneNumber()).orElse("");
-
-            if (newForceLocal) {
-                lastCountryCallingCode = getCountryCallingCode();
-                setPhoneNumber(localPhone);
-            } else if (lastCountryCallingCode != null) {
-                String prefix = String.valueOf(lastCountryCallingCode.phonePrefix());
-                if ((lastCountryCallingCode.countryCode() + localPhone).equals(prefix)) {
-                    setPhoneNumber(prefix);
-                } else {
-                    setPhoneNumber(lastCountryCallingCode.countryCode() + localPhone);
-                }
-            } else {
-                setPhoneNumber(localPhone);
-            }
-        }
-    };
-
-    /**
-     * @return Flag to force the user to enter a local number. This will change the default behaviour of the control and
-     * will have a local number on the {@link #phoneNumberProperty() phone number}.
-     */
-    public final BooleanProperty forceLocalNumberProperty() {
-        return forceLocalNumber;
-    }
-
-    public final boolean isForceLocalNumber() {
-        return forceLocalNumberProperty().get();
-    }
-
-    public final void setForceLocalNumber(boolean forceLocalNumber) {
-        forceLocalNumberProperty().set(forceLocalNumber);
-    }
-
     private final ObjectProperty<Callback<CountryCallingCode, Node>> countryCodeViewFactory = new SimpleObjectProperty<>(this, "countryCodeViewFactory");
 
     /**
@@ -361,84 +265,57 @@ public class PhoneNumberField extends Control {
         countryCodeViewFactoryProperty().set(countryCodeViewFactory);
     }
 
-    private final ReadOnlyStringWrapper mask = new ReadOnlyStringWrapper(this, "mask") {
-        @Override
-        public void set(String newMask) {
-            if (newMask != null && !newMask.isEmpty()) {
-                for (char c : newMask.toCharArray()) {
-                    if (!DEFAULT_MASK_SUPPORTED_CHARS.contains(c)) {
-                        throw new IllegalArgumentException("Mask contains unsupported character: " + c);
-                    }
-                }
-            }
-            super.set(newMask);
-        }
-    };
+    private final ObjectProperty<Callback<String, String>> phoneNumberFormatter = new SimpleObjectProperty<>(this, "phoneNumberFormatter");
 
     /**
-     * @return The mask the control will use to format phone numbers.
+     * @return The formatter that is used to format the phone number.
+     * This gets called whenever the {@link #phoneNumberProperty() phone number} changes.
      */
-    public final ReadOnlyStringProperty maskProperty() {
-        return mask.getReadOnlyProperty();
+    public final ObjectProperty<Callback<String, String>> phoneNumberFormatterProperty() {
+        return phoneNumberFormatter;
     }
 
-    public final String getMask() {
-        return mask.get();
+    public Callback<String, String> getPhoneNumberFormatter() {
+        return phoneNumberFormatterProperty().get();
     }
 
-    private void setMask(String mask) {
-        this.mask.set(mask);
+    public void setPhoneNumberFormatter(Callback<String, String> phoneNumberFormatter) {
+        phoneNumberFormatterProperty().set(phoneNumberFormatter);
     }
 
-    private final ReadOnlyStringWrapper maskRemaining = new ReadOnlyStringWrapper(this, "maskRemaining", getMask());
+    private final ObjectProperty<Callback<String, Boolean>> phoneNumberValidator = new SimpleObjectProperty<>(this, "phoneNumberValidator");
 
     /**
-     * @return A property that allows to know whether the mask is completed or not.
+     * @return The validator used to determine whether the phone number is valid or not.
+     * This gets called everytime the {@link #phoneNumberProperty() phone number} changes.
      */
-    public final ReadOnlyStringProperty maskRemainingProperty() {
-        return maskRemaining.getReadOnlyProperty();
+    public final ObjectProperty<Callback<String, Boolean>> phoneNumberValidatorProperty() {
+        return phoneNumberValidator;
     }
 
-    public final String getMaskRemaining() {
-        return maskRemaining.get();
+    public Callback<String, Boolean> getPhoneNumberValidator() {
+        return phoneNumberValidatorProperty().get();
     }
 
-    private void setMaskRemaining(String mask) {
-        this.maskRemaining.set(mask);
+    public void setPhoneNumberValidator(Callback<String, Boolean> phoneNumberValidator) {
+        phoneNumberValidatorProperty().set(phoneNumberValidator);
     }
 
-    private final ObjectProperty<Callback<PhoneNumberField, String>> maskProvider = new SimpleObjectProperty<>(this, "maskProvider");
-    /**
-     * @return The callback used to determine the mask. This gets called everytime the phone number changes.
-     */
-    public final ObjectProperty<Callback<PhoneNumberField, String>> maskProviderProperty() {
-        return maskProvider;
-    }
-
-    public final Callback<PhoneNumberField, String> getMaskProvider() {
-        return maskProviderProperty().get();
-    }
-
-    public final void setMaskProvider(Callback<PhoneNumberField, String> maskProvider) {
-        maskProviderProperty().set(maskProvider);
-    }
-
-    private final BooleanProperty strictMode = new SimpleBooleanProperty(this, "strictMode");
+    private final ReadOnlyBooleanWrapper valid = new ReadOnlyBooleanWrapper(this, "valid");
 
     /**
-     * @return Flag that defines the control to strictly follow the mask or not.  If set to true, the control will not allow to enter
-     * numbers beyond the mask definition. If set to false, you can enter any number of digits even thought they overrun the mask.
+     * @return Read only property that indicates whether the phone number is valid or not.
      */
-    public final BooleanProperty strictModeProperty() {
-        return strictMode;
+    public final ReadOnlyBooleanWrapper validProperty() {
+        return valid;
     }
 
-    public final boolean isStrictMode() {
-        return strictModeProperty().get();
+    public final boolean isValid() {
+        return valid.get();
     }
 
-    public final void setStrictMode(boolean strictMode) {
-        strictModeProperty().set(strictMode);
+    private void setValid(boolean valid) {
+        this.valid.set(valid);
     }
 
     /**
@@ -464,13 +341,6 @@ public class PhoneNumberField extends Control {
         String iso2Code();
 
         /**
-         * This is arbitrary adopted by the control, but can be replaced by using a custom {@link #maskProviderProperty() maskProvider}.
-         * @param localPhoneNumber The current local phone number.
-         * @return The default mask defined for the country according to the given local phone number.
-         */
-        String mask(String localPhoneNumber);
-
-        /**
          * @return The first area code if there is any in the country.
          */
         default Integer defaultAreaCode() {
@@ -480,14 +350,14 @@ public class PhoneNumberField extends Control {
         /**
          * @return The concatenation of country code and {@link #defaultAreaCode() default area code}.
          */
-        default int phonePrefix() {
-            StringBuilder value = new StringBuilder();
+        default String phonePrefix() {
+            StringBuilder value = new StringBuilder("+");
             value.append(countryCode());
             Integer defaultAreaCode = defaultAreaCode();
             if (defaultAreaCode != null) {
                 value.append(defaultAreaCode);
             }
-            return Integer.parseInt(value.toString());
+            return value.toString();
         }
 
         /**
@@ -541,8 +411,8 @@ public class PhoneNumberField extends Control {
             CHAD(235, "TD"),
             CHILE(56, "CL"),
             CHINA(86, "CN"),
-            CHRISTMAS_ISLAND(61, "CX", "(_____) _____", 89164),
-            COCOS_ISLANDS(61, "CC", "(_____) _____", 89162),
+            CHRISTMAS_ISLAND(61, "CX", 89164),
+            COCOS_ISLANDS(61, "CC", 89162),
             COLOMBIA(57, "CO"),
             COMOROS(269, "KM"),
             CONGO(242, "CG"),
@@ -576,15 +446,7 @@ public class PhoneNumberField extends Control {
             GABON(241, "GA"),
             GAMBIA(220, "GM"),
             GEORGIA(995, "GE"),
-            GERMANY(49, "DE", "(___) ______") {
-                @Override
-                public String mask(String localPhoneNumber) {
-                    if (localPhoneNumber != null && localPhoneNumber.startsWith("01522")) {
-                        return "(_____) _______";
-                    }
-                    return super.mask(localPhoneNumber);
-                }
-            },
+            GERMANY(49, "DE"),
             GHANA(233, "GH"),
             GIBRALTAR(350, "GI"),
             GREECE(30, "GR"),
@@ -593,7 +455,7 @@ public class PhoneNumberField extends Control {
             GUADELOUPE(590, "GP"),
             GUAM(1, "GU", 671),
             GUATEMALA(502, "GT"),
-            GUERNSEY(44, "GG", "(____) ______", 1481, 7781, 7839, 7911),
+            GUERNSEY(44, "GG", 1481, 7781, 7839, 7911),
             GUINEA(224, "GN"),
             GUINEA_BISSAU(245, "GW"),
             GUYANA(592, "GY"),
@@ -607,14 +469,14 @@ public class PhoneNumberField extends Control {
             IRAN(98, "IR"),
             IRAQ(964, "IQ"),
             IRELAND(353, "IE"),
-            ISLE_OF_MAN(44, "IM", "(____) ______", 1624, 7524, 7624, 7924),
+            ISLE_OF_MAN(44, "IM", 1624, 7524, 7624, 7924),
             ISRAEL(972, "IL"),
             ITALY(39, "IT"),
             IVORY_COAST(225, "CI"),
             JAMAICA(1, "JM", 658, 876),
             JAN_MAYEN(47, "SJ", 79),
             JAPAN(81, "JP"),
-            JERSEY(44, "JE", "(____) ______", 1534),
+            JERSEY(44, "JE", 1534),
             JORDAN(962, "JO"),
             KAZAKHSTAN(7, "KZ", 6, 7),
             KENYA(254, "KE"),
@@ -751,17 +613,11 @@ public class PhoneNumberField extends Control {
             private final int countryCode;
             private final String iso2Code;
             private final int[] areaCodes;
-            private final String localNumberMask;
 
             Defaults(int countryCode, String iso2Code, int... areaCodes) {
-                this(countryCode, iso2Code, DEFAULT_MASK, areaCodes);
-            }
-
-            Defaults(int countryCode, String iso2Code, String localNumberMask, int... areaCodes) {
                 this.countryCode = countryCode;
                 this.iso2Code = iso2Code;
                 this.areaCodes = Optional.ofNullable(areaCodes).orElse(new int[0]);
-                this.localNumberMask = localNumberMask;
             }
 
             @Override
@@ -779,10 +635,6 @@ public class PhoneNumberField extends Control {
                 return iso2Code;
             }
 
-            @Override
-            public String mask(String localPhoneNumber) {
-                return localNumberMask;
-            }
         }
 
     }
@@ -820,6 +672,7 @@ public class PhoneNumberField extends Control {
         private CountryCallingCode inferBestMatch(List<CountryCallingCode> matchingCodes) {
             CountryCallingCode code;
             if (matchingCodes.size() > 1) {
+
                 // TODO Here there will be some ambiguity since two countries have same score
                 // we need some sort of logic here, for now using the last one.
                 code = matchingCodes.get(matchingCodes.size() - 1);
@@ -832,6 +685,10 @@ public class PhoneNumberField extends Control {
 
         private CountryCallingCodeScore rankAndSplit(CountryCallingCode code, String phoneNumber) {
             if (phoneNumber != null && !phoneNumber.isEmpty()) {
+                if (phoneNumber.startsWith("+")) {
+                    phoneNumber = phoneNumber.substring(1);
+                }
+
                 String countryPrefix = String.valueOf(code.countryCode());
 
                 if (code.areaCodes().length == 0) {
@@ -845,9 +702,7 @@ public class PhoneNumberField extends Control {
                     for (int areaCode : code.areaCodes()) {
                         String areaCodePrefix = countryPrefix + areaCode;
                         if (phoneNumber.startsWith(areaCodePrefix)) {
-                            String localNumber = phoneNumber.length() > areaCodePrefix.length() ?
-                                phoneNumber.substring(areaCodePrefix.length()) :
-                                phoneNumber.substring(countryPrefix.length());
+                            String localNumber = phoneNumber.substring(countryPrefix.length());
                             return new CountryCallingCodeScore(2, localNumber);
                         }
                     }
@@ -859,6 +714,9 @@ public class PhoneNumberField extends Control {
 
         private boolean isInvalid(String newPhoneNumber) {
             if (newPhoneNumber != null && !newPhoneNumber.isEmpty()) {
+                if (newPhoneNumber.startsWith("+")) {
+                    newPhoneNumber = newPhoneNumber.substring(1);
+                }
                 for (char c : newPhoneNumber.toCharArray()) {
                     if (!Character.isDigit(c)) {
                         return true;
@@ -867,6 +725,7 @@ public class PhoneNumberField extends Control {
             }
             return false;
         }
+
     }
 
     /**
@@ -874,33 +733,21 @@ public class PhoneNumberField extends Control {
      */
     private final class PhoneNumberFormatter implements UnaryOperator<TextFormatter.Change> {
 
-        private boolean selfUpdate;
-        private final TextField textField;
-
-        private PhoneNumberFormatter() {
-            textField = new TextField();
+        private PhoneNumberFormatter(TextField textField) {
             textField.setTextFormatter(new TextFormatter<>(this));
-            formattedLocalPhoneNumber.bind(textField.textProperty());
+            textField.addEventHandler(KeyEvent.KEY_PRESSED, e -> {
+                if (e.getCode() == KeyCode.BACK_SPACE
+                    && (textField.getText() == null || textField.getText().isEmpty())
+                    && getCountryCallingCode() != null) {
+
+                    // Clear up the country code if the user deletes the entire text
+                    setPhoneNumber(null);
+                    e.consume();
+                }
+            });
         }
 
-        private void setFormattedLocalPhoneNumber(String newPhoneNumber) {
-            if (selfUpdate) {
-                // Ignore when I'm the one who initiated the update
-                return;
-            }
-            try {
-                selfUpdate = true;
-                String formattedPhoneNumber = doFormat(newPhoneNumber);
-                textField.setText(formattedPhoneNumber);
-                textField.positionCaret(formattedPhoneNumber == null ? 0 : formattedPhoneNumber.length());
-            } finally {
-                selfUpdate = false;
-            }
-        }
-
-        private boolean isUnMasked() {
-            return getMask() == null || getMask().isEmpty();
-        }
+        private boolean selfUpdate;
 
         @Override
         public TextFormatter.Change apply(TextFormatter.Change change) {
@@ -913,21 +760,23 @@ public class PhoneNumberField extends Control {
 
                 if (change.isAdded() || change.isReplaced()) {
                     String text = change.getText();
+                    if (getCountryCallingCode() == null && text.equals("+") && change.getControlNewText().equals("+")) {
+                        // Only allow if it is the first character
+                        return change;
+                    }
+
                     if (!text.matches("[0-9]+")) {
                         return null;
                     }
                 }
 
-                if (getCountryCallingCode() == null && !isForceLocalNumber()) {
-                    resolveCountryCode(change);
-                } else {
-                    if (change.isAdded()) {
-                        numbersAdded(change);
-                    } else if (change.isReplaced()) {
-                        // Not allowed
-                        change = null;
-                    } else if (change.isDeleted()) {
-                        numbersRemoved(change);
+                if (change.isContentChange()) {
+                    if (getCountryCallingCode() == null) {
+                        resolveCountryCode(change);
+                    } else {
+                        String newNationalNumber = undoFormat(change.getControlNewText());
+                        String newPhoneNumber = "+" + getCountryCallingCode().countryCode() + newNationalNumber;
+                        setPhoneNumber(newPhoneNumber);
                     }
                 }
 
@@ -942,9 +791,7 @@ public class PhoneNumberField extends Control {
             String newText = change.getControlNewText();
             PhoneNumber number = parser.call(newText);
             if (number != null) {
-                // TODO this won't ever pick a sub country, since always top countries are resolved first.  Try to
-                // find a solution for this
-                setPhoneNumber(String.valueOf(number.countryCallingCode.phonePrefix()));
+                setCountryCallingCode(number.countryCallingCode);
                 textField.setText(number.localPhoneNumber);
                 change.setText("");
                 change.setCaretPosition(0);
@@ -953,76 +800,34 @@ public class PhoneNumberField extends Control {
             }
         }
 
-        private void numbersAdded(TextFormatter.Change change) {
-            if (isUnMasked()) {
-                setPhoneNumber(undoFormat(change.getControlNewText()));
-                return;
+        private String doFormat(String localPhoneNumber) {
+            if (localPhoneNumber == null || localPhoneNumber.isEmpty()) {
+                return localPhoneNumber == null ? "" : localPhoneNumber;
             }
 
-            String remainingMask = Optional.ofNullable(getMaskRemaining()).orElse("");
-            String originalText = change.getText();
-            StringBuilder newText = new StringBuilder();
-
-            for (char number : originalText.toCharArray()) {
-                if (remainingMask.isEmpty()) {
-                    if (isStrictMode()) {
-                        break;
-                    }
-                    newText.append(number);
-                    continue;
-                }
-
-                char maskC;
-
-                do {
-                    maskC = remainingMask.charAt(0);
-                    remainingMask = remainingMask.substring(1);
-                    if (maskC == DEFAULT_MASK_DIGIT_CHAR) {
-                        newText.append(number);
-                    } else {
-                        newText.append(maskC);
-                    }
-                } while (maskC != DEFAULT_MASK_DIGIT_CHAR && !remainingMask.isEmpty());
+            if (getPhoneNumberFormatter() == null) {
+                return localPhoneNumber;
             }
 
-            change.setText(newText.toString());
-            change.setCaretPosition(change.getCaretPosition() + newText.length() - originalText.length());
-            change.setAnchor(change.getCaretPosition());
+            String prefix = "+" + getCountryCallingCode().countryCode();
+            String formattedNumber = getPhoneNumberFormatter().call(prefix + localPhoneNumber);
 
-            setPhoneNumber(undoFormat(change.getControlNewText()));
+            if (formattedNumber == null) {
+                return localPhoneNumber;
+            }
+
+            if (!formattedNumber.startsWith(prefix)) {
+                return localPhoneNumber;
+            }
+
+            return formattedNumber.substring(prefix.length());
         }
 
-        private void numbersRemoved(TextFormatter.Change change) {
-            if (isUnMasked()) {
-                setPhoneNumber(undoFormat(change.getControlNewText()));
-                return;
-            }
-
-            String originalText = change.getControlNewText();
-            int start = change.getRangeStart();
-
-            for (int i = originalText.length() - 1; i >= 0; i--) {
-                char c = originalText.charAt(i);
-                if (!Character.isDigit(c)) {
-                    start--;
-                } else {
-                    break;
-                }
-            }
-
-            change.setRange(start, change.getRangeEnd());
-            setPhoneNumber(undoFormat(change.getControlNewText()));
-        }
-
-        private String undoFormat(String formattedPhoneNumber) {
+        private String undoFormat(String formattedLocalPhoneNumber) {
             StringBuilder phoneNumber = new StringBuilder();
 
-            if (getCountryCallingCode() != null) {
-                phoneNumber.append(getCountryCallingCode().countryCode());
-            }
-
-            if (formattedPhoneNumber != null && !formattedPhoneNumber.isEmpty()) {
-                for (char  c: formattedPhoneNumber.toCharArray()) {
+            if (formattedLocalPhoneNumber != null && !formattedLocalPhoneNumber.isEmpty()) {
+                for (char  c: formattedLocalPhoneNumber.toCharArray()) {
                     if (Character.isDigit(c)) {
                         phoneNumber.append(c);
                     }
@@ -1032,42 +837,19 @@ public class PhoneNumberField extends Control {
             return phoneNumber.toString();
         }
 
-        private String doFormat(String phoneNumber) {
-            if (phoneNumber == null || phoneNumber.isEmpty()) {
-                return phoneNumber;
+        private void setFormattedLocalPhoneNumber(String localPhoneNumber) {
+            if (selfUpdate) {
+                return; // Ignore when I'm the one who initiated the update
             }
 
-            if (isUnMasked()) {
-                return phoneNumber;
+            try {
+                selfUpdate = true;
+                String formattedPhoneNumber = doFormat(localPhoneNumber);
+                textField.setText(formattedPhoneNumber);
+                textField.positionCaret(formattedPhoneNumber.length());
+            } finally {
+                selfUpdate = false;
             }
-
-            StringBuilder formattedPhoneNumber = new StringBuilder();
-
-            for (char maskChar : getMask().toCharArray()) {
-                if (phoneNumber.isEmpty()) {
-                    if (maskChar == DEFAULT_MASK_DIGIT_CHAR) {
-                        break;
-                    }
-                    formattedPhoneNumber.append(maskChar);
-
-                } else {
-                    char numberChar = phoneNumber.charAt(0);
-
-                    if (maskChar == DEFAULT_MASK_DIGIT_CHAR) {
-                        formattedPhoneNumber.append(numberChar);
-                        phoneNumber = phoneNumber.substring(1);
-                    } else {
-                        formattedPhoneNumber.append(maskChar);
-                        if (numberChar == maskChar) {
-                            phoneNumber = phoneNumber.substring(1);
-                        }
-                    }
-                }
-            }
-
-            formattedPhoneNumber.append(phoneNumber);
-
-            return formattedPhoneNumber.toString();
         }
 
     }
